collapse `app-shell-new`/`web-new` into `app-shell`/`web`
