--- shell/app-shell/.htaccess
@@ -1,9 +1,0 @@
-<filesMatch "\.(html|htm|js|css|json|manifest*)$">
-  FileETag None
-  <ifModule mod_headers.c>
-     Header unset ETag
-     Header set Cache-Control "max-age=0, no-cache, no-store, must-revalidate"
-     Header set Pragma "no-cache"
-     Header set Expires "Wed, 11 Jan 1984 05:00:00 GMT"
-  </ifModule>
-</filesMatch>
