+++ shell/app-shell/artifacts/arc-types.manifest
