+++ shell/app-shell/artifacts/identity-types.manifest
