+++ shell/app-shell/artifacts/launcher.manifest
