+++ shell/app-shell/artifacts/profile.manifest
