+++ shell/app-shell/elements/arc-steps.js
@@ -29,7 +29,7 @@ class ArcSteps extends Xen.Base {
       // `plan` has been instantiated into host, record it into `steps`
       this._addStep(plan, plans.generations, state.steps, state.applied);
     }
-    if (steps && plans) {
+    if (state.steps && plans) {
       // find a step from `steps` that correspondes to a plan in `plans` but hasn't been `applied`
       this._providePlanStep(plans, state.steps, state.applied);
     }
@@ -80,6 +80,8 @@ class ArcSteps extends Xen.Base {
       // do something smarter than literal matching anyway...
       // Find all mapped handles to be remembered.
       // Store as string, as we'll only use it to find exact matches later. (String is easier to compare)
+      // TODO(wkorman): Rename `views` below to `handles` which may
+      // necessitate revising the launcher.
       let mappedHandles = plan.handles
         .filter(v => (v.fate == 'map') && (v.id.substr(0, 7) == 'shared:'))
         .map(v => v.id)
