+++ shell/app-shell/elements/icon.svg.js
