+++ shell/app-shell/elements/menu-panel.js
