+++ shell/app-shell/elements/pop-menu.js
