+++ shell/app-shell/elements/share-picker.js
