+++ shell/app-shell/elements/shell-handles.js
