+++ shell/app-shell/elements/user-picker.js
