+++ shell/app-shell/icons.css.js
