+++ shell/apps/chrome-extension/index.html
@@ -2,44 +2,18 @@
 <html lang="en">
 <head>
   <title>Arcs</title>
+
   <meta charset="utf-8">
-  <link rel="manifest" href="manifest.json"/>
   <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
+
+  <link rel="manifest" href="manifest.json"/>
   <link rel="shortcut icon" href="../logo_64.png">
+  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Google+Sans">
+  <link rel="stylesheet" href="index.css">
 
-  <link href="https://fonts.googleapis.com/css?family=Google+Sans:400,500,700" rel="stylesheet">
-
-  <style>
-    body {
-      font-family: sans-serif;
-      margin: 0;
-      overflow-x: hidden;
-      overflow-y: scroll;
-    }
-    @font-face {
-      font-family: 'Material Icons';
-      font-style: normal;
-      font-weight: 400;
-      src: url(https://fonts.gstatic.com/s/materialicons/v30/2fcrYFNaTjcS6g4U3t-Y5ZjZjT5FdEJ140U2DJYC3mY.woff2) format('woff2');
-    }
-  </style>
-
-  <script>
-    window.shellPath = "../..";
-  </script>
-
-  <!-- TODO(wkorman): Ideally move Firebase script load to arc-store.js.
-      Currently the FireBase library is not ES Module compatible.
-      See also https://github.com/firebase/firebase-js-sdk/issues/100 -->
-  <script src="https://www.gstatic.com/firebasejs/4.2.0/firebase.js"></script>
-  <!-- firebase config is required before loading runtime -->
-  <script type="module" src="../../app-shell/firebase-config.js"></script>
-
-  <!-- load runtime (from build [build] or source [source])-->
-  <script type="module" src="../../build/ArcsLib.js"></script>
-  <!--<script type="module" src="../../source/ArcsLib.js"></script>-->
-
+  <script type="module" src="index.debug.js"></script>
   <script type="module" src="extension-app-shell.js"></script>
+  
 </head>
 <body>
