--- shell/apps/web-new/index.debug.html
@@ -1,23 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-
-  <title>Arcs</title>
-
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
-
-  <link rel="manifest" href="manifest.json"/>
-  <link rel="shortcut icon" href="../logo_64.png">
-  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Google+Sans">
-  <link rel="stylesheet" href="index.css">
-  
-  <script type="module" src="index.debug.js"></script>
-
-</head>
-<body>
-
-  <app-shell></app-shell>
-
-</body>
-</html>
