--- shell/apps/web-new/manifest.json
@@ -1,25 +1,0 @@
-{
-  "name": "Arcs",
-  "short_name": "Arcs",
-  "start_url": "./",
-  "display": "standalone",
-  "theme_color": "#ffffff",
-  "background_color": "#ffffff",
-  "icons": [
-    {
-      "src": "../logo_64.png",
-      "sizes": "64x64",
-      "type": "image/png"
-    },
-    {
-      "src": "../logo_128.png",
-      "sizes": "128x128",
-      "type": "image/png"
-    },
-    {
-      "src": "../logo_256.png",
-      "sizes": "256x256",
-      "type": "image/png"
-    }
-  ]
-}
