+++ shell/apps/web/config.js
