+++ shell/apps/web/index.js
@@ -3,8 +3,8 @@ import './config.js';
 // firebase runtime (customized by sjmiles@ for import-ability)
 import '../../components/firebase.4.2.0.js';
 // firebase config is required before loading arcs runtime
-import '../../app-shell-new/firebase-config.js';
+import '../../app-shell/firebase-config.js';
 // arcs runtime
-import '../../source/ArcsLib.js';
+import '../../build/ArcsLib.js';
 // elements
-import '../../app-shell-new/app-shell.js';
+import '../../app-shell/app-shell.js';
