--- shell/apps/web/launcher.manifest
@@ -1,8 +1,0 @@
-// Copyright (c) 2017 Google Inc. All rights reserved.
-// This code may only be used under the BSD style license found at
-// http://polymer.github.io/LICENSE.txt
-// Code distributed by Google as part of this project is also
-// subject to an additional IP rights grant found at
-// http://polymer.github.io/PATENTS.txt
-
-import 'https://$cdn/artifacts/Arcs/Arcs.recipes'
